3/6 phase 4

+++ server/src/main/java/Service/LogoutServices.java
@@ -9,10 +9,14 @@ import Model.AuthData;
 public class LogoutServices {
     public void logoutServices(AuthData authData) throws DataAccessException {
         AuthDAO authDAO = new SQLAuthDao();
+        if (!authDAO.findAuthToken(authData)){
+            throw new DataAccessException("Error: unauthorized");
+        }
         try{
             authDAO.deleteAuthtoken(authData);
-        } catch (DataAccessException e){
-            throw e;
+        }
+        catch (DataAccessException exception){
+            throw exception;
         }
     }
 }
